Add basic command to set and show cpu state.
This change add the ability to show the current cpu state
and set indivigual cpus online or offline

This change adds the ability to show libvirt domain xmls
by name or uuid.

This change alters the list --domains command to remove the
vcpu info.

This change add defusexml as a currently unused dep which will
be consumed to safely parse domain xmls in the future.