extend python version testing
This change re-adds support for python 3.6 this mainly
involes removing the user of thelru cache decorator and
adding some extra deps to backport dataclass support and
package metadata parseing.

This change adds supprot for python 3.10 which already worked
but should now be tested in ci.

This change extends the tox github action to run the tests tox
env on both python 3.6 and 3.10

This change also fixes some minor url issues with the setup.cfg