add ablity to build tcib image using a podman container.
This change adds some basic tool script to build tcib images.
The tcib-image.sh script builds a contianer with tcib and
copies in the build-images.sc script.
build-images build all vendor images with tcib in skip build mode
which generates the docker files but does not execute the buildah
build.

the run.sh script puts all of this together to have tcib generate
teh dockerfile then it invokes podman on the host to buidl the
resulting contaienrs and tags them.

this avoid the fact that nested buildah build inside a podman
contaienr do not work with the triplo-client container as a base.

finally this correct some bugs in the vendor image to ensure
arbiterd is installed system wide instead of via a user install.